update MessageServiceImpl et IMessageService

+++ src/main/java/fr/erickfranco/cv_api/services/serviceinter/IMessageService.java
@@ -11,10 +11,10 @@ public interface IMessageService {
 
     List<Message> findAll();
 
-    Message findById(Integer id);
+    Message findById(Long id);
 
     Message save(Message message);
 
-    public void delete(Integer id);
+    public void delete(Long id);
 
 }
